// src/components/Dashboard/Board/Freight/FreightItem.types.ts
export type FreightType = {
	freightID: string;
	postTime: Date;
	weight: number;
	length: number;
	status: string;
	supplier: string;
	sender: string;
	orderer: string;
	deliverer: string;
	origin: string;
	destination: string;
	distance: string;
	pickupDate: string;
	pickupTimeStart: string;
	pickupTimeEnd: string;
	deliveryDate: string;
	deliveryTimeStart: string;
	deliveryTimeEnd: string;
	purchasePrice: number;
	orderPrice: number;
	priceMargin: number;
	commodity: string;
	contact: string;
	comments: string[] | never[];
};